projects/company-summary/pyproject.toml
[tool.poetry]
name = "company-summary"
version = "0.1.0"
description = ""
authors = ["yurukatsu <yuruyuru.kaihatsu@gmail.com>"]
readme = "README.md"
packages = [{include = "modules", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
arxiv = "^2.0.0"
aws = "^0.2.5"
awswrangler = "^3.4.2"
boto3 = "^1.33.7"
duckduckgo-search = "^3.9.9"
faiss-cpu = "^1.7.4"
google-api-python-client = "^2.109.0"
langchain = "^0.0.345"
newsapi-python = "^0.2.7"
openai = "^1.3.7"
pypdf = "^3.17.1"
pymupdf = "^1.23.7"
python-dotenv = "^1.0.0"
ragas = "^0.0.21"
tiktoken = "^0.5.2"
tqdm = "^4.66.1"
voyageai = "^0.1.4"
wikipedia = "^1.4.0"
yfinance = "^0.2.32"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
notebook = "^7.0.6"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
